Job_Seekers/requirements.txt
streamlit==1.30.0
mysql-connector-python==8.3.0
protobuf>=3.20,<5
streamlit-lottie
